Add search

+++ webapp/views.py
@@ -1,12 +1,18 @@
 from django.shortcuts import render, redirect, get_object_or_404
 
 from webapp.models import Article
-from webapp.forms import ArticleForm
+from webapp.forms import ArticleForm, ArticleSearch
 
 
 def index_view(request):
-    articles = Article.objects.filter(status="active").order_by("-created_at")
-    context = {"articles": articles}
+    form = ArticleSearch()
+
+    if request.method == 'GET' and 'authorName' in request.GET:
+        authorName = request.GET['authorName']
+        articles = Article.objects.filter(authorName=authorName)
+    else:
+        articles = Article.objects.filter(status="active").order_by("-created_at")
+    context = {"articles": articles, "form": form}
     return render(request, "index.html", context)
 
 
@@ -25,27 +31,31 @@ def create_article(request):
         return render(request, "create.html", {"form": form})
 
 
-# def update_article(request, pk):
-#     article = get_object_or_404(Article, pk=pk)
-#     if request.method == "GET":
-#         form = ArticleForm(initial={
-#             "title": article.title,
-#             "author": article.author,
-#             "content": article.content
-#         })
-#         return render(request, "update.html", {"form": form})
-#     else:
-#         form = ArticleForm(data=request.POST)
-#         if form.is_valid():
-#             article.title = form.cleaned_data.get("title")
-#             article.author = form.cleaned_data.get("author")
-#             article.content = form.cleaned_data.get("content")
-#             article.save()
-#             return redirect("article_view", pk=article.pk)
-#         return render(request, "update.html", {"form": form})
+def update_article(request, pk):
+    article = get_object_or_404(Article, pk=pk)
+    if request.method == "GET":
+        form = ArticleForm(initial={
+            "authorName": article.authorName,
+            "authorEmail": article.authorEmail,
+            "content": article.content
+        })
+        return render(request, "update.html", {"form": form})
+    else:
+        form = ArticleForm(data=request.POST)
+        if form.is_valid():
+            article.authorName = form.cleaned_data.get("authorName")
+            article.authorEmail = form.cleaned_data.get("authorEmail")
+            article.content = form.cleaned_data.get("content")
+            article.save()
+            return redirect("index")
+        return render(request, "update.html", {"form": form})
 
 
 def delete_article(request, pk):
     article = get_object_or_404(Article, pk=pk)
-    article.delete()
-    return redirect("index")
+    if request.method == "GET":
+        pass
+        return render(request, "delete.html", {"article": article})
+    else:
+        article.delete()
+        return redirect("index")
